requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
google-generativeai
python-dotenv
fpdf2
email-validator
requests